requirements.txt
streamlit
langchain
openai
chromadb
pymupdf
selenium
beautifulsoup4
python-dotenv


# python -m spacy download en_core_web_sm
#pip install tika
